refactor: remove build-main code fences
