+++ app/scripts/controllers/metametrics-controller.ts
@@ -80,15 +80,12 @@ import {
   type EndTraceRequest,
   type TraceCallback,
 } from '../../../shared/lib/trace';
-
-///: BEGIN:ONLY_INCLUDE_IF(build-main)
 import { ENVIRONMENT } from '../../../development/build/constants';
-///: END:ONLY_INCLUDE_IF
-
 import { KeyringType } from '../../../shared/constants/keyring';
 import type { captureException } from '../../../shared/lib/sentry';
 import type { FlattenedBackgroundStateProxy } from '../../../shared/types';
 import { getTokensControllerAllTokens } from '../../../shared/modules/selectors/assets-migration';
+import { isMain } from '../../../shared/lib/build-types';
 import type {
   PreferencesControllerGetStateAction,
   PreferencesControllerStateChangeEvent,
@@ -946,15 +943,14 @@ export default class MetaMetricsController extends BaseController<
       this.setMarketingCampaignCookieId(null);
     }
 
-    ///: BEGIN:ONLY_INCLUDE_IF(build-main)
     if (
+      isMain() &&
       this.#environment !== ENVIRONMENT.DEVELOPMENT &&
       metaMetricsId !== null &&
       participateInMetaMetrics !== null
     ) {
       this.updateExtensionUninstallUrl(participateInMetaMetrics, metaMetricsId);
     }
-    ///: END:ONLY_INCLUDE_IF
 
     return metaMetricsId;
   }
