+++ ui/components/multichain/account-overview/common.ts
@@ -1,6 +1,4 @@
 export type AccountOverviewCommonProps = {
   setBasicFunctionalityModalOpen: () => void;
-  ///: BEGIN:ONLY_INCLUDE_IF(build-main)
   onSupportLinkClick: () => void;
-  ///: END:ONLY_INCLUDE_IF
 };
