+++ ui/pages/onboarding-flow/onboarding-flow-switch/onboarding-flow-switch.tsx
@@ -32,7 +32,7 @@ import {
   isExperimental,
   isFlask,
   isMain,
-} from '../../../helpers/utils/build-types';
+} from '../../../../shared/lib/build-types';
 
 // TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31860
 // eslint-disable-next-line @typescript-eslint/naming-convention
